+++ relay-cabi/src/processing.rs
@@ -273,6 +273,7 @@ pub unsafe extern "C" fn relay_store_normalizer_normalize_event(
         measurements: None,
         normalize_spans: config.normalize_spans,
         replay_id: config.replay_id,
+        span_allowed_hosts: &[], // only supported in relay
     };
     normalize_event(&mut event, &normalization_config);