+++ relay-event-normalization/src/normalize/span/tag_extraction.rs
@@ -16,7 +16,7 @@ use relay_event_schema::protocol::{
 use relay_protocol::{Annotated, Empty, Value};
 use sqlparser::ast::Visit;
 use sqlparser::ast::{ObjectName, Visitor};
-use url::Url;
+use url::{Host, Url};
 
 use crate::span::description::{
     concatenate_host_and_port, scrub_domain_name, scrub_span_description,
@@ -176,13 +176,22 @@ impl std::fmt::Display for RenderBlockingStatus {
 /// Wrapper for [`extract_span_tags`].
 ///
 /// Tags longer than `max_tag_value_size` bytes will be truncated.
-pub(crate) fn extract_span_tags_from_event(event: &mut Event, max_tag_value_size: usize) {
+pub(crate) fn extract_span_tags_from_event(
+    event: &mut Event,
+    max_tag_value_size: usize,
+    http_scrubbing_allow_list: &[Host],
+) {
     // Temporarily take ownership to pass both an event reference and a mutable span reference to `extract_span_tags`.
     let mut spans = std::mem::take(&mut event.spans);
     let Some(spans_vec) = spans.value_mut() else {
         return;
     };
-    extract_span_tags(event, spans_vec.as_mut_slice(), max_tag_value_size);
+    extract_span_tags(
+        event,
+        spans_vec.as_mut_slice(),
+        max_tag_value_size,
+        http_scrubbing_allow_list,
+    );
 
     event.spans = spans;
 }
@@ -190,7 +199,12 @@ pub(crate) fn extract_span_tags_from_event(event: &mut Event, max_tag_value_size
 /// Extracts tags and measurements from event and spans and materializes them into the spans.
 ///
 /// Tags longer than `max_tag_value_size` bytes will be truncated.
-pub fn extract_span_tags(event: &Event, spans: &mut [Annotated<Span>], max_tag_value_size: usize) {
+pub fn extract_span_tags(
+    event: &Event,
+    spans: &mut [Annotated<Span>],
+    max_tag_value_size: usize,
+    span_allowed_hosts: &[Host],
+) {
     // TODO: To prevent differences between metrics and payloads, we should not extract tags here
     // when they have already been extracted by a downstream relay.
     let shared_tags = extract_shared_tags(event);
@@ -208,7 +222,15 @@ pub fn extract_span_tags(event: &Event, spans: &mut [Annotated<Span>], max_tag_v
             continue;
         };
 
-        let tags = extract_tags(span, max_tag_value_size, ttid, ttfd, is_mobile, start_type);
+        let tags = extract_tags(
+            span,
+            max_tag_value_size,
+            ttid,
+            ttfd,
+            is_mobile,
+            start_type,
+            span_allowed_hosts,
+        );
 
         span.sentry_tags = Annotated::new(
             shared_tags
@@ -461,6 +483,7 @@ pub fn extract_tags(
     full_display: Option<Timestamp>,
     is_mobile: bool,
     start_type: Option<&str>,
+    span_allowed_hosts: &[Host],
 ) -> BTreeMap<SpanTagKey, String> {
     let mut span_tags: BTreeMap<SpanTagKey, String> = BTreeMap::new();
 
@@ -487,7 +510,7 @@ pub fn extract_tags(
             span_tags.insert(SpanTagKey::Category, category.to_owned());
         }
 
-        let (scrubbed_description, parsed_sql) = scrub_span_description(span);
+        let (scrubbed_description, parsed_sql) = scrub_span_description(span, span_allowed_hosts);
 
         let action = match (category, span_op.as_str(), &scrubbed_description) {
             (Some("http"), _, _) => span
@@ -1451,7 +1474,7 @@ LIMIT 1
             .into_value()
             .unwrap();
 
-        extract_span_tags_from_event(&mut event, 200);
+        extract_span_tags_from_event(&mut event, 200, &[]);
 
         let spans = event.spans.value().unwrap();
 
@@ -1513,7 +1536,7 @@ LIMIT 1
             .into_value()
             .unwrap();
 
-        extract_span_tags_from_event(&mut event, 200);
+        extract_span_tags_from_event(&mut event, 200, &[]);
 
         let span = &event.spans.value().unwrap()[0];
 
@@ -1631,7 +1654,7 @@ LIMIT 1
             .into_value()
             .unwrap();
 
-        extract_span_tags_from_event(&mut event, 200);
+        extract_span_tags_from_event(&mut event, 200, &[]);
 
         let span_1 = &event.spans.value().unwrap()[0];
         let span_2 = &event.spans.value().unwrap()[1];
@@ -1689,7 +1712,7 @@ LIMIT 1
             .into_value()
             .unwrap();
 
-        extract_span_tags_from_event(&mut event, 200);
+        extract_span_tags_from_event(&mut event, 200, &[]);
 
         let span = &event
             .spans
@@ -1803,7 +1826,7 @@ LIMIT 1
             .into_value()
             .unwrap();
 
-        extract_span_tags_from_event(&mut event, 200);
+        extract_span_tags_from_event(&mut event, 200, &[]);
 
         let span_1 = &event.spans.value().unwrap()[0];
         let span_2 = &event.spans.value().unwrap()[1];
@@ -1926,7 +1949,7 @@ LIMIT 1
             .into_value()
             .unwrap();
 
-        extract_span_tags_from_event(&mut event, 200);
+        extract_span_tags_from_event(&mut event, 200, &[]);
 
         let span_1 = &event.spans.value().unwrap()[0];
         let span_2 = &event.spans.value().unwrap()[1];
@@ -2034,7 +2057,7 @@ LIMIT 1
             .into_value()
             .unwrap();
 
-        extract_span_tags_from_event(&mut event, 200);
+        extract_span_tags_from_event(&mut event, 200, &[]);
 
         let span = &event.spans.value().unwrap()[0];
 
@@ -2091,7 +2114,7 @@ LIMIT 1
             .into_value()
             .unwrap();
 
-        extract_span_tags_from_event(&mut event, 200);
+        extract_span_tags_from_event(&mut event, 200, &[]);
 
         let span = &event.spans.value().unwrap()[0];
         let tags = span.value().unwrap().sentry_tags.value().unwrap();
@@ -2119,7 +2142,7 @@ LIMIT 1
             .unwrap()
             .into_value()
             .unwrap();
-        let tags = extract_tags(&span, 200, None, None, false, None);
+        let tags = extract_tags(&span, 200, None, None, false, None, &[]);
 
         assert_eq!(
             tags.get(&SpanTagKey::BrowserName),
@@ -2159,7 +2182,7 @@ LIMIT 1
             .into_value()
             .unwrap();
 
-        extract_span_tags_from_event(&mut event, 200);
+        extract_span_tags_from_event(&mut event, 200, &[]);
 
         let span = &event.spans.value().unwrap()[0];
         let tags = span.value().unwrap().sentry_tags.value().unwrap();
@@ -2190,7 +2213,7 @@ LIMIT 1
             .unwrap()
             .into_value()
             .unwrap();
-        let tags = extract_tags(&span, 200, None, None, false, None);
+        let tags = extract_tags(&span, 200, None, None, false, None, &[]);
 
         assert_eq!(
             tags.get(&SpanTagKey::MessagingDestinationName),
@@ -2315,7 +2338,7 @@ LIMIT 1
             .into_value()
             .unwrap();
 
-        extract_span_tags_from_event(&mut event, 200);
+        extract_span_tags_from_event(&mut event, 200, &[]);
 
         let span = &event.spans.value().unwrap()[0];
         let tags = span.value().unwrap().sentry_tags.value().unwrap();
@@ -2413,7 +2436,7 @@ LIMIT 1
             .unwrap();
         span.description.set_value(Some(description.into()));
 
-        extract_tags(&span, 200, None, None, false, None)
+        extract_tags(&span, 200, None, None, false, None, &[])
     }
 
     #[test]